AC-265950: Updated code snippets, removed workflow related code, removed snippets for Activity 4 linking InvoiceNbr, and added a db script based on the T230 lesson revamp.
