+++ Customization/T230/CodeSnippets/Step3.2/RSSVRepairPriceMaint.cs
@@ -44,21 +44,22 @@ namespace PhoneRepairShop
             View DefaultWarranty;
         #endregion
 
-        private static void ValidatePrices(RSSVRepairPrice repairPriceItem)
+		////////// The added code
+	    private static void ValidatePrices(RSSVRepairPrice repairPriceItem)
         {
             // Create an instance of the RSSVRepairPriceMaint graph.
             var repairPrices = PXGraph.CreateInstance<RSSVRepairPriceMaint>();
             repairPrices.RepairPrices.Current = repairPriceItem;
 
             // Set a delay to mimic connecting to an external service to validate the 
-            // repair item prices.
+			// repair item prices.
             // In a real world scenario, you would connect to an actual external service and 
-            // make an API request to validate the prices for the repair items.
+			// make an API request to validate the prices for the repair items.
             Thread.Sleep(3000);
 
             // Update the Price Validated field for each repair item on the Repair Items tab:
             // Here we are assuming that the validation was successful from the external service 
-            // and are setting IsPriceValidated to true for each repair item.     
+			// and are setting IsPriceValidated to true for each repair item.     
             foreach (RSSVRepairItem item in repairPrices.RepairItems.Select())
             {
                 // Set IsPriceValidated to true for each repair item.
@@ -69,34 +70,7 @@ namespace PhoneRepairShop
             // Trigger the Save action to save the changes stored in the cache to the database.
             repairPrices.Save.Press();
         }
-
-        #region Actions
-        public PXAction<RSSVRepairPrice> ValidateItemPrices;
-        [PXButton(DisplayOnMainToolbar = false)]
-        [PXUIField(DisplayName = "Validate Prices", Enabled = true)]
-        protected virtual IEnumerable validateItemPrices(PXAdapter adapter)
-        {
-            // Populate a local list variable.
-            List<RSSVRepairPrice> list = new List<RSSVRepairPrice>();
-            foreach (RSSVRepairPrice repairItemPrice in adapter.Get<RSSVRepairPrice>())
-            {
-                list.Add(repairItemPrice);
-            }
-
-            // Trigger the Save action to save changes in the database.
-            Actions.PressSave();
-
-            var repairPriceItem = RepairPrices.Current;
-
-            PXLongOperation.StartOperation(this, delegate () {
-                ValidatePrices(repairPriceItem);
-            });
-
-            // Return the local list variable.
-            return list;
-        }
-        #endregion
-
+		////////// The end of added code
         #region Event Handlers
         //Update the price and repair item type when the inventory ID of
         //the repair item is updated.
