+++ Customization/T230/PhoneRepairShop_Code/PhoneRepairShop_Code/PhoneRepairShop_Code.csproj
@@ -81,7 +81,6 @@
     <Compile Include="RSSVSetupMaint.cs" />
     <Compile Include="DAC\RSSVSetup.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Workflow\RSSVWorkOrderWorkflow.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Bcl.AsyncInterfaces">
