+++ Customization/T230/SourceFiles/DBScripts/T230_AddColumn_RSSVRepairItem.sql
@@ -1,0 +1,3 @@
+ALTER TABLE [SmartFix_T230].[dbo].[RSSVRepairItem]
+ADD IsPriceValidated bit NOT NULL
+DEFAULT (0)
